pyproject.toml
[tool.poetry]
name = "biblioterra"
version = "0.1.0"
description = ""
authors = ["lamarcke <cassiolamarcksilvafreitas@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.79.0"
uvicorn = "^0.17.6"
grab-fork-from-libgen = "^3.0.2"
aioredis = "^2.0.1"
gunicorn = "^20.1.0"
slowapi = "^0.1.5"
passlib = "^1.7.4"
python-dotenv = "^0.20.0"
python-multipart = "^0.0.5"
dnspython = "^2.2.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
motor = "^3.0.0"
fastapi-mail = "^1.0.9"
httpx = "^0.22.0"
aiomysql = "^0.1.1"
hurry-filesize = "^0.9"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
